update my_set method

+++ application/controllers/Action.php
@@ -106,7 +106,7 @@ class Action extends MY_Controller {
 			$child = scandir($path ."/".$v);
 			foreach($child as $val) {
 				
-				if($val != '.' && $val != '..') {
+				if($val != '.' && $val != '..' &&  $val != '.keep') {
 					$counter++;
 					unlink($path . "/{$v}/".$val);
 				}
